Fix crash when 16-byte stack alignment in needed
- don't call any C or SDL2 function directly from assembly code (wrapped in Wrapper.c),
- use __attribute__((force_align_arg_pointer)) for every function called from assembly code,
- allow to use "$CC" environment variable for compiler.

On Linux it prevents crash when one of system libraries needs 16-byte stack alignment due to SSE instructions.
This also allows the game to run (still not playable) on OS X.

I'm not using "-mstackrealign", because it doesn't fix the crash on GCC (fixes crash only on Clang).

This patch can insignificantly reduce the performance.
The stack realignment can be disabled in compilation script (remove -DSTACK_REALIGN).

Closes: #7