Using SDL_HAPTIC_STEERING_AXIS for other effects (#98)
Some of the effects was nothing loading correctly under windows as the game tries to send polar coordinates to a single axis device (racing wheel) and this is not supported on windows.
This fix those effects and the game now have rumble when driving on different surfaces.
Collision effect still forces the wheel to move to left (CCW). I can't find a fix for it.
Spring effect is still broken. There's a commented out code to apply a "force center spring" effect if desired. This also needs more work.
Note: Windows users should not enable `useHapticPolar` on ini file.